# docs/OpenAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Users API",
    "description": "Users API to Orbi",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
    "/v1/health": {
      "get": {
        "operationId": "healthCheck",
        "tags": [
          "Health check"
        ],
        "responses": {
          "200": {
            "description": "OK!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }

      }

    },
    "/v1/users/{id}": {
      "get": {
        "operationId": "getUser",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateUser",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/CreateUser"
        },
        "responses": {
          "200": {
            "description": "OK!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users": {
      "post": {
        "operationId": "createUser",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/CreateUser"
        },
        "responses": {
          "201": {
            "description": "Success scenario!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Yael"
          },
          "age": {
            "type": "integer",
            "example": 23
          },
          "email": {
            "type": "string",
            "example": "contacto@yael.mx"
          }
        }
      }
    },
    "requestBodies": {
      "CreateUser": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      }
    }
  }
}